move src/components/dashboard to its own domain
