+++ src/routes.js
@@ -5,7 +5,7 @@ var React = require('react'),
   Wrapper = require('./containers/wrapper'),
   Login = require('./auth/Login'),
   Logout = require('./auth/Logout'),
-  Dashboard = require('./components/dashboard'),
+  Dashboard = require('./dashboard/dashboard'),
   Profile = require('./components/settings'),
   LastUserSelector = require('./components/lastUserSelector'),
   Authorize = require('./auth/Authorize');
